+++ classes/ListFilterForm.php
@@ -12,8 +12,10 @@
 
 namespace HeimrichHannot\FormHybridList;
 
+use Contao\Session;
 use HeimrichHannot\FormHybrid\Form;
 use HeimrichHannot\Haste\Util\Url;
+use HeimrichHannot\Request\Request;
 
 class ListFilterForm extends Form
 {
@@ -34,6 +36,19 @@ class ListFilterForm extends Form
         $objModule->filterHeadline           = is_array($arrHeadline) ? $arrHeadline['value'] : $arrHeadline;
         $objModule->filterHl                 = is_array($arrHeadline) ? $arrHeadline['unit'] : 'h1';
         $objModule->formHybridSkipValidation = true;
+        
+        // reset session if filter is reset
+        if(Request::getGet('reset'))
+        {
+            $this->resetSession(session_id().'_filter_'.$objModule->id);
+        }
+        
+        // update default values when filter is saved in session
+        if($objModule->saveFilterToSession && null !== ($sessionFilter = \Session::getInstance()->get(session_id().'_filter_'.$objModule->id)))
+        {
+            $objModule->formHybridAddDefaultValues = true;
+            $objModule->formHybridDefaultValues = $this->updateDefaultValues($objModule, $sessionFilter);
+        }
 
         parent::__construct($objModule);
     }
@@ -41,7 +56,7 @@ class ListFilterForm extends Form
     protected function onSubmitCallback(\DataContainer $dc)
     {
         $this->submission = $dc;
-
+        
         if (is_array($this->arrSubmitCallbacks) && !empty($this->arrSubmitCallbacks)) {
             foreach ($this->arrSubmitCallbacks as $arrCallback) {
                 if (is_array($arrCallback) && !empty($arrCallback)) {
@@ -49,6 +64,27 @@ class ListFilterForm extends Form
                 }
             }
         }
+        
+        if($this->objListModule->saveFilterToSession)
+        {
+            $this->saveFilterToSession();
+        }
+    }
+    
+    protected function saveFilterToSession()
+    {
+        $submission = $this->getSubmission(false)->row();
+        $module = $this->objListModule;
+        $sessionFilter = [];
+        
+        foreach(deserialize($module->customFilterFields,true) as $value)
+        {
+            $sessionFilter[$value] = $submission[$value];
+        }
+        
+        \Session::getInstance()->set(session_id().'_filter_'.$module->id, $sessionFilter);
+        
+        \Controller::redirect(\Haste\Util\Url::removeQueryString(array_merge(['FORM_SUBMIT'], array_keys($this->arrFields))));
     }
 
     protected function compile()
@@ -57,10 +93,12 @@ class ListFilterForm extends Form
 
     protected function generateResetFilterField()
     {
+        $url = $this->objListModule->saveFilterToSession ? \Haste\Util\Url::addQueryString('reset=true', Url::getCurrentUrlWithoutParameters())  : Url::getCurrentUrlWithoutParameters();
+        
         $arrData = [
             'inputType' => 'explanation',
             'eval'      => [
-                'text' => '<div class="form-group reset-filter"><a class="btn btn-default btn-lg" href="' . Url::getCurrentUrlWithoutParameters() . '"><span>'
+                'text' => '<div class="form-group reset-filter"><a class="btn btn-default btn-lg" href="' . $url . '"><span>'
                     . $GLOBALS['TL_LANG']['formhybrid_list'][FORMHYBRID_LIST_BUTTON_RESET_FILTER][0] . '</span></a></div>',
             ],
         ];
@@ -129,5 +167,46 @@ class ListFilterForm extends Form
                 $this->{$callback[0]}->{$callback[1]}($arrDca);
             }
         }
+    }
+    
+    /**
+     * update default values according to filter from session
+     *
+     * @param $module
+     * @param $session
+     *
+     * @return string
+     */
+    protected function updateDefaultValues($module, $session)
+    {
+        if(empty($customFields = deserialize($module->customFilterFields,true)))
+        {
+            return '';
+        }
+        
+        $defaults =  [];
+        
+        foreach($customFields as $field)
+        {
+            if('' == $session[$field])
+            {
+                continue;
+            }
+         
+            $defaults[] = [
+                'field' => $field,
+                'value' => $session[$field],
+                'label' => ''
+            ];
+        }
+        
+        $filterValues = array_merge(deserialize($module->formHybridDefaultValues,true),$defaults);
+        
+        return serialize($filterValues);
+    }
+    
+    public function resetSession($key)
+    {
+        Session::getInstance()->set($key,null);
     }
 }