+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [5.2.1] - 2018-03-05
+## [5.3.0] - 2018-03-06
 
-#### Changed
-- added minified version of js
+#### Added
+- optional freetext search field
+- save the filter parameter to session in `ListFilterForm::onSubmitCallback()` and set them as defaultValues when building the form in `ListFilterForm::__construct()`
+- freetext and proximity search in `ModuleList::applyDefaultValues()`
 
 ## [5.2.0] - 2018-02-09